+++ packages/lumx-vue/src/components/toolbar/index.ts
@@ -1,0 +1,3 @@
+export { default as Toolbar } from './Toolbar';
+export type { ToolbarProps } from './Toolbar';
+export { TOOLBAR_NAME } from './Toolbar';