+++ packages/site-demo/content/product/components/toolbar/vue/demo2.vue
@@ -1,0 +1,15 @@
+<template>
+    <Toolbar>
+        <Heading typography="title">Toolbar title</Heading>
+        <template #after>
+            <Button emphasis="low" :left-icon="mdiTranslate" :right-icon="mdiMenuDown">
+                Value
+            </Button>
+        </template>
+    </Toolbar>
+</template>
+
+<script setup lang="ts">
+import { mdiMenuDown, mdiTranslate } from '@lumx/icons';
+import { Button, Toolbar, Heading } from '@lumx/vue';
+</script>